SQL sessions consolidation
