+++ lib/msf/core/exploit/remote/mssql.rb
@@ -205,7 +205,7 @@ module Exploit::Remote::MSSQL
   # Issue a SQL query using the TDS protocol
   #
   def mssql_query(sqla, doprint=false, opts={})
-    @mssql_client.mssql_query(sqla, doprint, opts)
+    @mssql_client.query(sqla, doprint, opts)
   end
 
   #
