+++ lib/rex/post/mssql/ui/console/command_dispatcher/core.rb
@@ -1,6 +1,6 @@
 # -*- coding: binary -*-
 
-require 'rex/post/mssql'
+require 'rex/post/sql/ui/console/command_dispatcher/core'
 
 module Rex
   module Post
@@ -12,48 +12,8 @@ module Rex
         #
         ###
         class Console::CommandDispatcher::Core
-
+          include Rex::Post::Sql::Ui::Console::CommandDispatcher::Core
           include Rex::Post::MSSQL::Ui::Console::CommandDispatcher
-
-          #
-          # Initializes an instance of the core command set using the supplied session and client
-          # for interactivity.
-          #
-          # @param [Rex::Post::MSSQL::Ui::Console] console
-
-          #
-          # List of supported commands.
-          #
-          def commands
-            cmds = {
-              '?' => 'Help menu',
-              'background' => 'Backgrounds the current session',
-              'bg' => 'Alias for background',
-              'exit' => 'Terminate the MSSQL session',
-              'help' => 'Help menu',
-              'irb' => 'Open an interactive Ruby shell on the current session',
-              'pry' => 'Open the Pry debugger on the current session',
-              'sessions' => 'Quickly switch to another session'
-            }
-
-            reqs = {}
-
-            filter_commands(cmds, reqs)
-          end
-
-          #
-          # Core
-          #
-          def name
-            'Core'
-          end
-
-          def unknown_command(cmd, line)
-            status = super
-
-            status
-          end
-
         end
       end
     end
