+++ lib/rex/post/mysql/ui/console/command_dispatcher/core.rb
@@ -1,5 +1,7 @@
 # -*- coding: binary -*-
 
+require 'rex/post/sql/ui/console/command_dispatcher/core'
+
 module Rex
   module Post
     module MySQL
@@ -7,42 +9,8 @@ module Rex
 
         # Core MySQL client commands
         class Console::CommandDispatcher::Core
-
+          include Rex::Post::Sql::Ui::Console::CommandDispatcher::Core
           include Rex::Post::MySQL::Ui::Console::CommandDispatcher
-
-          # List of supported commands.
-          #
-          # @return [Hash{String->String}]
-          def commands
-            cmds = {
-              '?'                        => 'Help menu',
-              'background'               => 'Backgrounds the current session',
-              'bg'                       => 'Alias for background',
-              'exit'                     => 'Terminate the MySQL session',
-              'help'                     => 'Help menu',
-              'irb'                      => 'Open an interactive Ruby shell on the current session',
-              'pry'                      => 'Open the Pry debugger on the current session',
-              'sessions'                 => 'Quickly switch to another session',
-            }
-
-            reqs = {}
-
-            filter_commands(cmds, reqs)
-          end
-
-          # @return [String]
-          def name
-            'Core'
-          end
-
-          # @param [Object] cmd
-          # @param [Object] line
-          # @return [Symbol, nil]
-          def unknown_command(cmd, line)
-            status = super
-
-            status
-          end
         end
       end
     end
