+++ lib/rex/post/postgresql.rb
@@ -1,3 +1,4 @@
 # -*- coding: binary -*-
 
+require 'rex/post/sql'
 require 'rex/post/postgresql/ui'
