+++ lib/rex/post/sql.rb
@@ -1,0 +1,3 @@
+# -*- coding: binary -*-
+
+require 'rex/post/sql/ui'
