+++ lib/rex/post/sql/ui.rb
@@ -1,0 +1,3 @@
+# -*- coding: binary -*-
+
+require 'rex/post/sql/ui/console'
