+++ spec/lib/rex/post/mssql/ui/console/command_dispatcher/core_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
-require 'rex/post/mssql/ui/console/command_dispatcher/core'
+require 'rex/post/mssql'
 
 RSpec.describe Rex::Post::MSSQL::Ui::Console::CommandDispatcher::Core do
   let(:rstream) { instance_double(::Rex::Socket) }
